+++ components/netsuite/src/actions/records/createRecord.ts
@@ -1,14 +1,19 @@
-import { action } from "@prismatic-io/spectral";
+import { action, outputSchema } from "@prismatic-io/spectral";
 import { createClient } from "../../client";
 import { createCustomerExamplePayload } from "../../examplePayloads";
 import { createRecordInputs } from "../../inputs";
+import { createRecordOutputSchema } from "../../outputSchemas";
 import { parseLocationData } from "../../utils";
 export const createRecord = action({
   display: {
     label: "Create Record",
-    description: "Create record of specified type",
+    description: "Create record of specified type.",
   },
   inputs: createRecordInputs,
+  outputSchema: outputSchema({
+    type: "actionOutput",
+    schema: createRecordOutputSchema,
+  }),
   examplePayload: createCustomerExamplePayload,
   perform: async (context, params) => {
     const client = await createClient(