+++ components/netsuite/src/dataSources/selectRecord.ts
@@ -5,20 +5,20 @@ import type { Record } from "../types/Record";
 export const selectRecord = dataSource({
   display: {
     label: "Select Record",
-    description: "Select a record from a list of records",
+    description: "Select a record from a list of records.",
   },
   inputs: selectRecordInputs,
   dataSourceType: "picklist",
   perform: async (
     _context,
-    { connection, recordType, query, limit, offset, recordField },
+    { connection, recordType, query, pagination, recordField },
   ) => {
     const client = await createClient(connection, "record", false);
     const { data } = await client.get(`/${recordType}`, {
       params: {
         q: query,
-        limit: limit,
-        offset: offset,
+        limit: pagination.limit,
+        offset: pagination.offset,
       },
     });
     const records: Record[] = data.items;