+++ components/netsuite/src/dataSources/selectSuiteQl.ts
@@ -4,17 +4,14 @@ import { selectSuiteQlInputs } from "../inputs";
 export const selectSuiteQl = dataSource({
   display: {
     label: "Select SuiteQL",
-    description: "Execute a SuiteQL Query to create a picklist",
+    description: "Execute a SuiteQL query to create a picklist.",
   },
   inputs: selectSuiteQlInputs,
   dataSourceType: "picklist",
-  perform: async (
-    _context,
-    { connection, query, limit, offset, key, label },
-  ) => {
+  perform: async (_context, { connection, query, pagination, key, label }) => {
     const client = await createClient(connection, "query", false);
     const { data } = await client.post(
-      `/suiteql?limit=${limit}&offset=${offset}`,
+      `/suiteql?limit=${pagination.limit}&offset=${pagination.offset}`,
       {
         q: query,
       },